backend/requirements.txt
Flask==2.3.2
flask-cors==3.0.10
SQLAlchemy==1.4.49
psycopg2-binary==2.9.6
python-dotenv==1.0.1
requests==2.31.0
python-multipart==0.0.6
pypdf==3.13.0
python-docx==0.8.11
flask-caching==2.0.2
sentence-transformers==2.2.2
numpy==1.26.4
